user_service/authentication/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import UserProfile


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = "__all__"