user_service/requirements.txt
Django
djangorestframework
psycopg2-binary
requests
djangorestframework-simplejwt
gunicorn
psycopg2-binary
python-decouple